Metemos los Many y One para hacer las relaciones.

+++ src/main/java/com/iesvdc/acceso/zapateria/models/Telefono.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Telefono {
     @Id
-    private Integer numero;
+    private Long numero;
     private int codigoPais;
 }
